modified:   __init__.py 	renamed:    authx/__init__.py -> trustauthx/__init__.py 	renamed:    authx/authlite.py -> trustauthx/authlite.py
